requirements.txt
matplotlib==3.4.3
opencv_python==4.5.3.56
numpy==1.21.2
cvlib==0.2.6
streamlit==0.88.0
Pillow==8.3.2
